requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
jinja2==3.1.2
python-multipart==0.0.6

# 智能识别相关
cn2an==0.5.22
jieba==0.42.1

# 网络请求
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# HTML解析
beautifulsoup4==4.12.2

# HTML解析
beautifulsoup4==4.12.2

# 数据处理
pandas==2.1.4
numpy==1.25.2

# 文件处理
watchdog==3.0.0
pathlib2==2.3.7.post1

# 加密和安全
cryptography==42.0.8

# 异步处理
celery==5.3.4
redis==5.0.1

# 数据库
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# 云存储
boto3==1.34.0
alibabacloud-oss-sdk==0.1.1
tencentcloud-sdk-python==3.0.1024

# 音乐识别相关
pyacoustid==1.2.2
chromaprint==1.5.0
mutagen==1.47.0
musicbrainzngs==0.7.1

# 媒体处理相关
moviepy==1.0.3
Pillow==10.1.0

# MoviePilot PT站点管理增强
lxml==4.9.3
cloudscraper==1.2.71
beautifulsoup4==4.12.2
pyyaml==6.0.1

# 下载器支持
qbittorrent-api==0.4.4
transmission-rpc==4.2.0

# 媒体识别增强
themoviedb==1.0.0
imdbpy==2021.4.18

# 日志和监控
structlog==23.2.0
prometheus-client==0.19.0

# 配置管理
python-dotenv==1.0.0
pyyaml==6.0.1

# 网络和API
websockets==12.0
starlette==0.27.0

# 系统工具
psutil==5.9.6
humanize==4.8.0

# 测试和开发
pytest==7.4.3
pytest-asyncio==0.21.1
